Add docker.Image component
The docker.Image component adds a nice interface over building and pusshing an image to a Docker registry, and getting a handle to the resulting image name and digest.

This can be used with any Docker registry.  For example, to push to Docker Hub:

```javascript
const image = new docker.Image("mynginx", {
    imageName: "hekul/mynginx:v1",
    build: "./mynginx",
    registry: {
        name: "docker.io",
        username: dockerUsername,
        password: dockerPassword,
    },
});
export const dockerImageName = image.imageName;
```