Ensure we do always login to a registry when issuing multiple login requests to the same one. (#38)
Fixes issue where we were incorrectly skipping logging into a docker registry when doing a parallel update with many resources coming from the same registry. The problem stemmed from a previous PR where we tried to prevent logging into the same docker registry multiple times. That change was buggy in that it recorded we were logged in prior to actually executing the login. That meant that when one update was actually doing the login (asynchronously), another resource could then move forward, try to login, think it was logged in (even though the first hadn't completed) and then would continue on to further docker operations which would then fail.

This change now makes it so that we store around the actual promise tracking the real login. While that login is asynchonously computing, if any other login requests come along, they will then see and await that real login, meaning none of those registry operations make progress until the first login to that registry returns.