Fix two usability issues
1. Mark docker.Container as DeleteBeforeReplace. Since docker consumes
single-use resources such as ports on the local machine, it is very
easy to run into situations where CreateBeforeDelete is doomed to fail.
Conservatively marking docker.Container as DBR ensures that we always
delete containers before trying to create their replacements.

2. Fix the naming of 'networks_advanced' to 'networksAdvanced', fitting
in with the naming convention of the rest of the package.