Changing the Name of the docker Config resource to be ServiceConfig
Fixes: #140

We had `docker_config` being created as `Config` in the NodeJS, Go
and Python packages. This was actually colliding with the `Config`
module of each of the packages but was particularly breaking the
Python integration. We are going to accept that this is technically
a breaking change BUT that it's for the better since it actually
makes the Config module usable

The DotNet SDK was already using `ServiceConfig` so we made the
other SDKs similar to this