Add container registry examples (#243)
* Add container registry examples

This change adds container registry examples for building and
publishing new container images to provisioned repositories in AWS,
Azure, GCP, DigitalOcean, and Docker Hub, using the Docker Image
component. Examplesa are provided in C#, Go, JavaScript, Python,
and TypeScript. Additionally, a Kubernetes example for the Docker
Hub, including secrets, plus a GCP example that deploys to GCR and
consumes that image from GKE, are included.

Note that none of these are included in the testing yet. That comes
next (and will undoubtedly happen incrementally).

* Add a CHANGELOG entry