Enable use of  Docker configuration for push registry authentication
We've used Docker api auth type to authenticate a Buildkit session.
For pushing to the registry, we can use this as well, using the registry
servername provided in the pulumi program as the lookup for which
registry to authenticate against.
If the program or provider config supplies credentials (username and
password), we use those preferentially.