Adds logic to parse registry name from a fully qualified image name.
The new `getRegistryAddr` will supply the web address of the registry
server for the purpose of looking up the registry auth in the host
machine's docker config.
When a registryServer is not supplied, the `getRegistryAddr` function
will attempt to parse the registry server address from the fully
qualified image name.
It will handle some special case logic around
Dockerhub as well.