Set Registry Server Name (#465)
* Adds logic to parse registry name from a fully qualified image name.

The new `getRegistryAddr` will supply the web address of the registry
server for the purpose of looking up the registry auth in the host
machine's docker config.
When a registryServer is not supplied, the `getRegistryAddr` function
will attempt to parse the registry server address from the fully
qualified image name.
It will handle some special case logic around
Dockerhub as well.

* Do not require registry:server

With the logic atted to image.go, the user no longer needs to explicitly
set the registry server name, as it can be inferred from a fully
qualified registry image name.

* Re-generate schema

We updated a few versions, so a new schema is in order.

* Re-generate SDKs

* Add  prefix to registry name on user's behalf

* Add comments to clarify the structure of authConfig object

* code review: remove errors package and include link to docker default issue

* Use Distribution library to parst registry host name from image name

Uses the
https://github.com/distribution/distribution/blob/main/reference/reference.go
package to verify correct format of a fully qualified image name. Note
that this does not verify the registry itslef; only that the image name
has valid format.
Additionally, adds error message to clarify a fully qualified image name
is required in all cases, which should catch anyone trying to use
Dockerhub shortnames..

* remove extraneous TODO

* Add test for getRegistryAddrFromImage

Also improves the output for the error message, removing the print
statement and wrapping the error instead.