Remove `env` from schema
The `env` field was used to pass along a buildkit option.
This functionality has been replaced by `builderVersion`.
Additionally, an `ENV` field is common to use in Dockerfiles, and to
pass to containers as runtime flags, so using this name is confusing as
well. ENV vars do not get passed at build time.