Add Dockerfile to build context if passed from outside the build context (#542)
Breaking Change: 
If a user provides a Dockerfile, the path specified must be absolute, or relative to the _Pulumi_ context, not the Docker build context.
The `dockerfile` field remains optional, and if left blank, we assume traditional location of Dockerfile within the Docker build context.

Specifics:
- If a user does not specify a Dockerfile the provider defaults to "Dockerfile", relative to Docker build context. The tarball already contains all necessary build files.
- If a user declares a Dockerfile, it must be an absolute path or a path relative to the Pulumi.yaml's location in the file system. The provider does its best to detect if a Dockerfile path was declared in error, and sends a helpful error message to the user during `Check()`, i.e. during a `pulumi preview`. 
- When a Dockerfile is passed from outside the Docker build context, we create a separate Dockerfile context and add it to
the build tarball. Note that any COPY statements are now relative to the Docker context, and not the Pulumi program's filesystem location.


* Add Dockerfile to build context using the clibuild method

Testing this against context: app and dockerfile: Dockerfile, we see
 error hashing dockerfile "Dockerfile": could not open file Dockerfile: open app/Dockerfile: no such file or directory
Unsure if I installed the correct provider version. Retrying.

* First attempt at getting a relative Dockerfile so we can hash it properly

* Pass relDockerfile to hashContext

Currently fails wilt dockerfile frontend error.
relDockerfile needs to be passed to the build context tarball.

* This now errors with

Diagnostics:
  docker:index:Image (my-image):
    error: failed to compute cache key: "/app" not found: not found

But that just means the Dockerfile is passing the wrong folder - this is
an expected error.

* Remove debugs and cleanup

* fix linter; disable dotnet registry tests

* A blank Dockerfile gets tests to pass but reverses the logic for outside-context Dockerfiles again. We need to create different logic for hashing and not use the cli build library.

* Use separator as way to determine whether this dockerfile is relative to the build context or not

* Use filepath separator detection to determine if a Dockerfile context is
needed.

* unskip dotnet tests

* Refactor relative path getter function for testing purposes

* Add unit test for getting the relative path to demonstrate and cement assumptions

* lint: handle error

* Remove default values for Dockerfile and Context. Refine logic for
detecting external Dockerfile.

Return dockerfile and context as outputs in schema.
Return additional outputs from Create: build context and its relative Dockerfile
Pass the replaced Dockerfile to the image build options when necessary

Co-authored-by: Aaron Friel <friel@pulumi.com>

* Verify existence of Dockerfile in Check function

This allows for us to fail early, during preview, and send an error
message to a user who is providing an erroneous Dockerfile.

Co-authored-by: Aaron Friel <friel@pulumi.com>

* Add integration tests for new dockerfile logic

Co-authored-by: Aaron Friel <friel@pulumi.com>

* Re-generate schema and SDKs

* Add a test for a build left blank

* Pacify linter

---------

Co-authored-by: Aaron Friel <friel@pulumi.com>