Refactor context hash to avoid computing relative Dockerfile path
This fixes #578 by refactoring context hashing to not rely on a relative
Dockerfile path. This check was a little error-prone, because it didn't account
for forward slashes on Windows as a valid path separator.

```go
  if strings.Contains(dockerfile, string(filepath.Separator)) {
```

However, we don't need to compute the relative path for the Dockerfile if we
change `hashContext` to allow the Dockerfile to be specified with an absolute or
relative path to the project directory.

The most obvious test to include here is that if the Dockerfile is outside of
the build context, the context hash is now agnostic to the name of the
Dockerfile. This matches Docker's behavior, and a unit test was added
accordingly.