Use fsutil.Walk() to compute context digest
This removes the `patternmatcher` library and replaces our context walk with the
same library as used in Buildkit.

It turns out we *may* have already fixed #607 in #596, as the tests pass without
changes to `provider.go`. However, there is logic in `fsutil.Walk()` absent in
our previous implementation and it makes sense to limit the scope of the
provider to reusing the routine from the `fsutil` library.

Fixes #607