Prevent unknown inflection in build (#799)
A somewhat bizarre fix for #676 

### Problem

Let me try to explain what is going on in #676. The relevant snippet is
(same as in the new test):

```
const image = new docker.Image("my-image", {
    build: {
        context: "app",
        cacheFrom: {
            images: [imageName]
        },
    }
```

Where `imageName` depends on another resource R. In the problematic
scenario, the resource R is being replaced, which turns `imageName` into
unknown. Unfortunately, this turns the entire `build` object into a big
unknown instead of the desired effect of it keeping its shape with only
an unknown value inside `cacheFrom`.

Here is a gRPC log:

```
{"build":"04da6b54-80e4-46f7-96ec-b56ff0331ba9","imageName":"04da6b54-80e4-46f7-96ec-b56ff0331ba9","registry":{"password":"04da6b54-80e4-46f7-96ec-b56ff0331ba9","server":"04da6b54-80e4-46f7-96ec-b56ff0331ba9","username":"04da6b54-80e4-46f7-96ec-b56ff0331ba9"},"skipPush":false}
```

It turns out that `build` is unknown because of the current
implementation of Node.js codegen that sets the parent object as unknown
if that object needs a default value (see
[here](https://github.com/pulumi/pulumi-docker/blob/24161acd0433b061e3b5acfa128b9ddb599b965b/sdk/nodejs/image.ts#L177)).

We then discard unknowns
[here](https://github.com/pulumi/pulumi-docker/blob/24161acd0433b061e3b5acfa128b9ddb599b965b/provider/provider.go#L131)
and proceed with filling default build properties in
[`marshalBuildAndApplyDefaults`](https://github.com/pulumi/pulumi-docker/blob/24161acd0433b061e3b5acfa128b9ddb599b965b/provider/provider.go#L139),
which causes the issue because `Dockerfile` is not available at its
default location.

### Fix

We don't really need to set the default `builderVersion` value in the
schema. We assume the same default value in
[`marshalBuilder`](https://github.com/pulumi/pulumi-docker/blob/24161acd0433b061e3b5acfa128b9ddb599b965b/provider/image.go#L563)
and cover this scenario with a few unit tests
([example](https://github.com/pulumi/pulumi-docker/blob/24161acd0433b061e3b5acfa128b9ddb599b965b/provider/image_test.go#L55)).
Therefore, in this PR I chose to remove the default value from schema
and SDKs and, as a side effect, avoid turning `build` into an unknown.

### Test

I tested it manually and adjusted the ACR test into a two-step test that
exercises the scenario. You can see the test fail before my change
[here](https://github.com/pulumi/pulumi-docker/actions/runs/6510330782/job/17684951125#step:27:166).
Also added another test to validate #620.

Fix #676
Fix #620