Update Moby dependencies, address skipped builds (#828)
Adding additional logging to `sess.Run()` call here, and fixing
moby/moby#45935, as a prerequisite to debugging #812.

While looking into #812, when using a locally built provider, updates to
the Image resource would not result in a new image built and a
corresponding new `RepoDigest` output value. This was root caused to the
Moby issue below after adding a log statement following `sess.Run()` in
`provider/image.go`.
- https://github.com/moby/moby/issues/45935

However, the behavior seen while reproducing is *different* from our
issue. That local repro was too consistent, it occurred in over 20
consecutive builds following the repro steps at the bottom of this
commit.

After updating the Moby dependencies, the image build was no longer
skipped and I was unable to reproduce #812. Neither [Docker v23 release
notes](https://docs.docker.com/engine/release-notes/23.0/) nor the [git
diff from 23.0.1 to
23.0.7](https://github.com/moby/moby/compare/v23.0.1...v23.0.7) contain
any obvious changes that explain that.

The most likely conclusion here is:
1. There is still an unknown bug, likely a race condition, that can
trigger issue 812, and I was unsuccessful in reproducing it.
2. The Moby dependency update fixed an issue with the session dialing
that has the same symptoms as issue 812, which is fixed in the change to
`client/client.go` in the Moby diff above.

Repro
=====

The Pulumi program in `examples/multi-stage-build-go` was run repeatedly
to trigger #812 via:

```console
$ uuidgen -r > ./app/example-lockfile.json && pulumi up --skip-preview --diff
```

When the symptoms occurred, verbose logging indicated the image build
was skipped entirely - `docker.ImageBuild` did not run.