Add the ability to ignore changes to volatile provider configuration
The password and username of the registryAuth config can be volatile
for many cloud registries. For AWS ECR you need to exchange AWS
credentials for short lived tokens to authenticate to the registry.
This will lead to perma-diffs in the provider config.

This change adds the ability to ignore certain volatile fields
of the registryAuth configuration.